Skinned meshes now have their bones exposed as entities. Cloning entities is now different.

+++ engine/runtime/meta/ecs/entity.hpp
@@ -6,7 +6,7 @@
 namespace runtime
 {
 
-std::map<std::uint32_t, runtime::entity>& get_serialization_map();
+std::map<std::uint64_t, runtime::entity>& get_serialization_map();
 
 SAVE_EXTERN(entity);
 LOAD_EXTERN(entity);
